refactor: move vscode context out of shared
