+++ lib/shared/src/chat/commands/cody.json
