+++ lib/shared/src/chat/commands/index.ts
