+++ lib/shared/src/chat/commands/templates.ts
