+++ lib/shared/src/chat/commands/utils.test.ts
