+++ lib/shared/src/chat/recipes/custom-prompt.ts
@@ -1,33 +1,10 @@
-import { URI } from 'vscode-uri'
-
-import { CodebaseContext } from '../../codebase-context'
-import { ContextMessage } from '../../codebase-context/messages'
-import { ActiveTextEditorSelection, Editor } from '../../editor'
-import { MAX_HUMAN_INPUT_TOKENS, NUM_CODE_RESULTS, NUM_TEXT_RESULTS } from '../../prompt/constants'
+import { EditorContext } from '../../editor-context'
+import { MAX_HUMAN_INPUT_TOKENS } from '../../prompt/constants'
 import { truncateText } from '../../prompt/truncation'
-import { CodyPromptContext, defaultCodyPromptContext } from '../prompts'
-import {
-    extractTestType,
-    getHumanLLMText,
-    isOnlySelectionRequired,
-    newInteraction,
-    newInteractionWithError,
-} from '../prompts/utils'
-import {
-    getCurrentDirContext,
-    getCurrentFileContextFromEditorSelection,
-    getCurrentFileImportsContext,
-    getDirectoryFileListContext,
-    getEditorDirContext,
-    getEditorOpenTabsContext,
-    getFilePathContext,
-    getHumanDisplayTextWithFileName,
-    getPackageJsonContext,
-    getTerminalOutputContext,
-} from '../prompts/vscode-context'
+import { defaultCodyPromptContext } from '../commands'
+import { getHumanLLMText, isOnlySelectionRequired, newInteraction, newInteractionWithError } from '../commands/utils'
 import { Interaction } from '../transcript/interaction'
 
-import { getFileExtension, numResults } from './helpers'
 import { Recipe, RecipeContext, RecipeID } from './recipe'
 
 /** ======================================================
@@ -43,13 +20,16 @@ export class CustomPrompt implements Recipe {
      * The Interaction object contains messages from both the human and the assistant, as well as context information.
      */
     public async getInteraction(commandRunnerID: string, context: RecipeContext): Promise<Interaction | null> {
+        if (!context.editorContext) {
+            console.error('Context builder not available')
+            return null
+        }
+
         const command = context.editor.controllers?.command?.getCommand(commandRunnerID)
         if (!command) {
             const errorMessage = 'Invalid command -- command not found.'
             return newInteractionWithError(errorMessage)
         }
-
-        const workspaceRootUri = context.editor.getWorkspaceRootUri()
 
         const contextConfig = command?.context || defaultCodyPromptContext
         // If selection is required, ensure not to accept visible content as selection
@@ -71,116 +51,32 @@ export class CustomPrompt implements Recipe {
             return newInteractionWithError(errorMessage, commandName)
         }
 
-        // Add selection file name as display when available
-        const displayText = getHumanDisplayTextWithFileName(commandName, selection, workspaceRootUri)
-        const text = getHumanLLMText(promptText, selection?.fileName)
-
-        // Attach code selection to prompt text if only selection is needed as context
-        if (selection && isOnlySelectionRequired(contextConfig)) {
-            const contextMessages = Promise.resolve(getCurrentFileContextFromEditorSelection(selection))
-            return newInteraction({ text, displayText, contextMessages })
-        }
-
         const commandOutput = command.context?.output
 
+        const text = getHumanLLMText(promptText, selection?.fileName)
         const truncatedText = truncateText(text, MAX_HUMAN_INPUT_TOKENS)
-        const contextMessages = this.getContextMessages(
+
+        const editorContext = new EditorContext(
+            context.editorContext,
             truncatedText,
             context.editor,
             context.codebaseContext,
-            contextConfig,
             selection,
             commandOutput
         )
 
+        // Add selection file name as display when available
+        const displayText = editorContext.getHumanDisplayText(commandName)
+
+        // Attach code selection to prompt text if only selection is needed as context
+        if (selection && isOnlySelectionRequired(contextConfig)) {
+            const contextMessages = Promise.resolve(editorContext.getCurrentFileContextFromEditorSelection())
+            return newInteraction({ text, displayText, contextMessages })
+        }
+
+        // Create context messages for the command based on the context configuration
+        const contextMessages = editorContext.getContextMessages(contextConfig)
+
         return newInteraction({ text: truncatedText, displayText, contextMessages })
-    }
-
-    private async getContextMessages(
-        text: string,
-        editor: Editor,
-        codebaseContext: CodebaseContext,
-        promptContext: CodyPromptContext,
-        selection?: ActiveTextEditorSelection | null,
-        commandOutput?: string | null
-    ): Promise<ContextMessage[]> {
-        const contextMessages: ContextMessage[] = []
-        const workspaceRootUri = editor.getWorkspaceRootUri()
-        const isUnitTestRequest = extractTestType(text) === 'unit'
-
-        if (promptContext.none) {
-            return []
-        }
-
-        if (promptContext.codebase) {
-            const codebaseMessages = await codebaseContext.getContextMessages(text, numResults)
-            contextMessages.push(...codebaseMessages)
-        }
-        if (promptContext.openTabs) {
-            const openTabsMessages = await getEditorOpenTabsContext()
-            contextMessages.push(...openTabsMessages)
-        }
-        if (promptContext.currentDir) {
-            const currentDirMessages = await getCurrentDirContext(isUnitTestRequest)
-            contextMessages.push(...currentDirMessages)
-        }
-        if (!promptContext.directoryPath) {
-            if (promptContext.directoryPath) {
-                const dirMessages = await getEditorDirContext(promptContext.directoryPath, selection?.fileName)
-                contextMessages.push(...dirMessages)
-            }
-        }
-        if (!promptContext.filePath) {
-            if (promptContext.filePath) {
-                const fileMessages = await getFilePathContext(promptContext.filePath)
-                contextMessages.push(...fileMessages)
-            }
-        }
-
-        // Context for unit tests requests
-        if (isUnitTestRequest && contextMessages.length === 0) {
-            if (selection?.fileName) {
-                const importsContext = await this.getUnitTestContextMessages(selection, workspaceRootUri)
-                contextMessages.push(...importsContext)
-            }
-        }
-
-        if (promptContext.currentFile || promptContext.selection !== false) {
-            if (selection) {
-                const currentFileMessages = getCurrentFileContextFromEditorSelection(selection)
-                contextMessages.push(...currentFileMessages)
-            }
-        }
-        if (promptContext.command && commandOutput) {
-            const outputMessages = getTerminalOutputContext(commandOutput)
-            contextMessages.push(...outputMessages)
-        }
-        // Return sliced results
-        const maxResults = Math.floor((NUM_CODE_RESULTS + NUM_TEXT_RESULTS) / 2) * 2
-        return contextMessages.slice(-maxResults * 2)
-    }
-
-    private async getUnitTestContextMessages(
-        selection: ActiveTextEditorSelection,
-        workspaceRootUri?: URI | null
-    ): Promise<ContextMessage[]> {
-        const contextMessages: ContextMessage[] = []
-
-        if (workspaceRootUri) {
-            const rootFileNames = await getDirectoryFileListContext(workspaceRootUri, true)
-            contextMessages.push(...rootFileNames)
-        }
-        // Add package.json content only if files matches the ts/js extension regex
-        if (selection?.fileName && getFileExtension(selection?.fileName).match(/ts|js/)) {
-            const packageJson = await getPackageJsonContext(selection?.fileName)
-            contextMessages.push(...packageJson)
-        }
-        // Try adding import statements from current file as context
-        if (selection?.fileName) {
-            const importsContext = await getCurrentFileImportsContext()
-            contextMessages.push(...importsContext)
-        }
-
-        return contextMessages
     }
 }
