+++ lib/shared/src/chat/recipes/recipe.ts
@@ -1,5 +1,6 @@
 import { CodebaseContext } from '../../codebase-context'
 import { Editor } from '../../editor'
+import { EditorContextMessages } from '../../editor-context'
 import { IntentDetector } from '../../intent-detector'
 import { BotResponseMultiplexer } from '../bot-response-multiplexer'
 import { Interaction } from '../transcript/interaction'
@@ -11,6 +12,7 @@ export interface RecipeContext {
     codebaseContext: CodebaseContext
     responseMultiplexer: BotResponseMultiplexer
     firstInteraction: boolean
+    editorContext?: EditorContextMessages
 }
 
 export type RecipeID =
