+++ lib/shared/src/index.ts
@@ -5,4 +5,4 @@ export { ChatButton, ChatMessage } from './chat/transcript/messages'
 export { renderCodyMarkdown } from './chat/markdown'
 export { basename, pluralize, isDefined, dedupeWith } from './common'
 export { ContextFile, PreciseContext } from './codebase-context/messages'
-export { CodyPrompt } from './chat/prompts'
+export { CodyPrompt } from './chat/commands'
