+++ lib/shared/tsconfig.json
@@ -6,6 +6,6 @@
     "outDir": "dist",
     "lib": ["ESNext", "DOM.Iterable"],
   },
-  "include": ["src", "src/chat/prompts/cody.json"],
+  "include": ["src", "src/chat/commands/cody.json"],
   "exclude": ["dist"],
 }
